break out limit into own function
