+++ packages/swapper/src/swappers/thorchain/utils/ethereum/utils/getThorTxData.ts
@@ -3,6 +3,7 @@ import { fromAssetId } from '@shapeshiftoss/caip'
 
 import { SwapError, SwapErrorTypes } from '../../../../../api'
 import { InboundResponse, ThorchainSwapperDeps } from '../../../types'
+import { getLimit } from '../../getLimit/getLimit'
 import { makeSwapMemo } from '../../makeSwapMemo/makeSwapMemo'
 import { thorService } from '../../thorService'
 import { deposit } from '../routerCalldata'
@@ -51,7 +52,7 @@ export const getThorTxInfo: GetBtcThorTxInfo = async ({
         details: { inboundAddresses }
       })
 
-    const memo = await makeSwapMemo({
+    const limit = await getLimit({
       buyAssetId: buyAsset.assetId,
       destinationAddress,
       sellAmount,
@@ -60,6 +61,12 @@ export const getThorTxInfo: GetBtcThorTxInfo = async ({
       slippageTolerance,
       deps,
       tradeFee
+    })
+
+    const memo = await makeSwapMemo({
+      buyAssetId: buyAsset.assetId,
+      destinationAddress,
+      limit
     })
     const data = await deposit(
       router,
